requirements.txt
numpy
pandas
matplotlib
Pillow
stepic
cryptography
streamlit
